Delinkify reference to dakosta.org from index.html
Technically the link points to a different page, but
that different page ends up being a redirection to
this page again. It felt wrong to have that, even if
the link would bring you to where the canonical website
were actually hosted.

It also probably improves the experience by adhering
to the implicit standard that the link that brought you
to page x is not a link when you get to page x.